refactor(layout): move health/ to serve/health/ (T124.3.1)
